pyproject.toml
[project]
name = "telegram-sales-bot"
version = "0.1.0"
description = "Telegram bot for sales"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.2",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.41",
]
